+++ CHANGELOG.md
@@ -19,6 +19,11 @@ Master
 end 6.5
 --------
 
+- Core: Validate theme and language. Don't echo given language string to fix XSS issue
+- Core: Set "samesite=Strict" on all cookies
+- Notes: Improved list style
+- Core: Custom text fields default to empty string instead of null
+
 12-01-2021 6.4.210
 - Core: Fixed constraint error in database check